Review fixes

+++ backend/src/PetFamily.Application/SpeciesHandlers/Commands/Delete/DeleteSpeciesCommandHandler.cs
@@ -4,6 +4,7 @@ using Microsoft.Extensions.Logging;
 using PetFamily.Application.Abstractions;
 using PetFamily.Application.Database;
 using PetFamily.Application.Extensions;
+using PetFamily.Application.VolunteersHandlers;
 using PetFamily.Domain.Shared;
 using PetFamily.Domain.Shared.IDs;
 
@@ -14,17 +15,20 @@ public class DeleteSpeciesCommandHandler : ICommandHandler<DeleteSpeciesCommand>
     private readonly ILogger<DeleteSpeciesCommandHandler> _logger;
     private readonly IUnitOfWork _unitOfWork;
     private readonly IValidator<DeleteSpeciesCommand> _validator;
+    private readonly IVolunteerRepository _volunteerRepository;
     private readonly ISpeciesRepository _repository;
 
     public DeleteSpeciesCommandHandler(
         ILogger<DeleteSpeciesCommandHandler> logger,
         IUnitOfWork unitOfWork,
         IValidator<DeleteSpeciesCommand> validator,
+        IVolunteerRepository volunteerRepository,
         ISpeciesRepository repository)
     {
         _logger = logger;
         _unitOfWork = unitOfWork;
         _validator = validator;
+        _volunteerRepository = volunteerRepository;
         _repository = repository;
     }
 
@@ -43,6 +47,13 @@ public class DeleteSpeciesCommandHandler : ICommandHandler<DeleteSpeciesCommand>
         if(speciesResult.IsFailure)
             return speciesResult.Error.ToErrorList();
 
+        var volunteers = await _volunteerRepository.GetAll(cancellationToken);
+        var isSpeciesActive = volunteers.Any(v =>
+            v.Pets.Any(p => p.SpeciesDetails.SpeciesId == speciesId));
+
+        if(isSpeciesActive)
+            return Errors.General.ValueIsInvalid(nameof(speciesId)).ToErrorList();
+        
         _repository.Delete(speciesResult.Value);
 
         await _unitOfWork.SaveChanges(cancellationToken);
